requirements.txt
# Nutze pip install, um die benötigten python Pakete zu erlangen
# Gib folgendes in deiner CMD ein : pip install -r requirements.txt

Flask>=1.0.0,<1.1
chatterbot==0.8.7
# SQLAlchemy>=1.3.0,<1.4
Pillow>=5.0
pytz>=2018
pyperclip